pyproject.toml
[tool.poetry]
name = "gebrauchtwaffen-aggregator"
version = "0.1.0"
description = "Swiss used firearms aggregator - personal web app for Raspberry Pi"
authors = ["Löli"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
jinja2 = "^3.1.3"
python-dotenv = "^1.0.0"
httpx = "^0.26.0"
alembic = "^1.13.1"
beautifulsoup4 = "^4.12.0"
lxml = "^5.1.0"
python-multipart = "^0.0.22"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
